requirements_route_optimization.txt
Flask==2.3.2
requests==2.31.0
google-maps-routeoptimization==0.1.11
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0 